/trunk: cisstStereoVision - VS2008 x64 errors and warnings fixed
git-svn-id: https://svn.lcsr.jhu.edu/cisst/trunk@3106 0959eb1a-6956-4f60-947b-6092a4e22c97

+++ cisst/cisstStereoVision/code/svlDraw.cpp
@@ -278,7 +278,7 @@ void svlDraw::Poly(svlSampleImage* image,
 {
     if (!image || videoch >= image->GetVideoChannels()) return;
 
-    const unsigned int size = points.size();
+    const unsigned int size = static_cast<unsigned int>(points.size());
 
     if (size < 1) return;
     if (size == 1) Pixel(image, videoch, points[0], color);
